backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-socketio
python-multipart
websockets
bcrypt==3.2.2
passlib==1.7.4
python-jose[cryptography]
